# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LKPM-Flow API — companies, reports, data points, constraints, audit trail, and dashboards for LKPM (Laporan Kegiatan Penanaman Modal) reporting to OSS.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List reports
      parameters:
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
        - name: izinId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ReportStatus"
        - name: scale
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/BusinessScale"
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Create a report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a report with full detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateReport
      tags: [reports]
      summary: Update a report (status, narrative, gates)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /reports/{id}/approval:
    post:
      operationId: transitionApproval
      tags: [reports]
      summary: Advance the maker-checker-approver approval flow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalActionInput"
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "403":
          description: Not authorized for this approval step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Invalid transition for current status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    ApprovalStatus:
      type: string
      enum: [draft, submitted, reviewed, approved]
      description: Status alur persetujuan maker-checker-approver
    ApprovalActionInput:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [submit, review, approve, reject]
          description: >-
            submit = maker mengajukan; review = checker menyetujui tinjauan;
            approve = approver menyetujui akhir; reject = kembalikan ke draft
        note:
          type: string
          description: Catatan opsional untuk jejak audit
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    BusinessScale:
      type: string
      enum: [mikro, kecil, menengah, besar]
      description: Investment scale determining reporting frequency
    OperatingMode:
      type: string
      enum: [pendamping, hibrida, penuh]
      description: Working mode — UMK companion, hybrid, or full human-in-loop
...
    ReportStatus:
      type: string
      enum: [intake, collect, validate, draft, review, submit, monitor, archive]
      description: Pipeline stage of the report
    PeriodType:
      type: string
      enum: [triwulan, semester, tahunan]
    DataCategory:
      type: string
      enum: [investasi, tenaga_kerja, produksi, kewajiban]
      description: LKPM data component category
    DataStatus:
      type: string
      enum: [terverifikasi, perlu_verifikasi, estimasi]
      description: Verification status of a data point
...
    Report:
      type: object
      required:
        - id
        - izinId
        - companyId
        - companyName
        - projectName
        - scale
        - operatingMode
        - periodType
        - periodLabel
        - year
        - deadline
        - status
        - idIzin
        - createdAt
      properties:
        id:
          type: integer
        izinId:
          type: integer
        companyId:
          type: integer
        companyName:
          type: string
        projectName:
          type: ["string", "null"]
          description: Nama proyek dari Izin terkait
        scale:
          $ref: "#/components/schemas/BusinessScale"
        operatingMode:
          $ref: "#/components/schemas/OperatingMode"
        periodType:
          $ref: "#/components/schemas/PeriodType"
        periodLabel:
          type: string
          description: e.g. "TW II 2026" or "Semester I 2026"
        year:
          type: integer
        deadline:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/ReportStatus"
        idIzin:
          type: string
          description: Id Izin / NIB project identifier
        narrative:
          type: ["string", "null"]
          description: Overall constraint narrative for the report
        ossReceipt:
          type: ["string", "null"]
          description: OSS submission receipt reference
        makerName:
          type: ["string", "null"]
        checkerName:
          type: ["string", "null"]
        approverName:
          type: ["string", "null"]
        makerId:
          type: ["string", "null"]
        checkerId:
          type: ["string", "null"]
        approverId:
          type: ["string", "null"]
        approvalStatus:
          $ref: "#/components/schemas/ApprovalStatus"
        createdAt:
          type: string
          format: date-time
    ReportDetail:
      type: object
      required:
        - report
        - dataPoints
        - constraints
        - activities
      properties:
        report:
          $ref: "#/components/schemas/Report"
        dataPoints:
          type: array
          items:
            $ref: "#/components/schemas/DataPoint"
        constraints:
          type: array
          items:
            $ref: "#/components/schemas/Constraint"
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
    ReportInput:
      type: object
      required:
        - izinId
        - periodType
        - periodLabel
        - year
        - deadline
      properties:
        izinId:
          type: integer
        periodType:
          $ref: "#/components/schemas/PeriodType"
        periodLabel:
          type: string
          minLength: 1
        year:
          type: integer
        deadline:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/ReportStatus"
        narrative:
          type: string
        makerName:
          type: string
        checkerName:
          type: string
        approverName:
          type: string
    ReportUpdate:
      type: object
      properties:
        periodType:
          $ref: "#/components/schemas/PeriodType"
        periodLabel:
          type: string
          minLength: 1
        year:
          type: integer
        deadline:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/ReportStatus"
        narrative:
          type: string
        ossReceipt:
          type: string
        makerName:
          type: ["string", "null"]
        checkerName:
          type: ["string", "null"]
        approverName:
          type: ["string", "null"]
        makerId:
          type: ["string", "null"]
        checkerId:
          type: ["string", "null"]
        approverId:
          type: ["string", "null"]
...
    DataPoint:
      type: object
      required:
        - id
        - reportId
        - category
        - label
        - status
        - confidence
        - createdAt
      properties:
        id:
          type: integer
        reportId:
          type: integer
        category:
          $ref: "#/components/schemas/DataCategory"
        label:
          type: string
          description: e.g. "Realisasi modal tetap"
        fieldKey:
          type: ["string", "null"]
          description: >-
            Stable identifier binding this data point to an OSS catalog field
            (e.g. "inv_mesin"). When present, it is authoritative for
            catalog matching so label edits do not break the binding.
        value:
          type: ["number", "null"]
          description: Numeric realization value
        unit:
          type: ["string", "null"]
          description: e.g. "IDR", "orang"
        source:
          type: ["string", "null"]
          description: Provenance citation, e.g. "GL-2026-Q2, baris 88"
        status:
          $ref: "#/components/schemas/DataStatus"
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        attribution:
          type: ["string", "null"]
          description: Who/which agent produced this value
        createdAt:
          type: string
          format: date-time
...
    Constraint:
      type: object
      required:
        - id
        - reportId
        - issue
        - createdAt
      properties:
        id:
          type: integer
        reportId:
          type: integer
        issue:
          type: string
        followUp:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Activity:
      type: object
      required:
        - id
        - reportId
        - action
        - actor
        - createdAt
      properties:
        id:
          type: integer
        reportId:
          type: integer
        action:
          type: string
          description: e.g. "Gate G3 lolos", "Data divalidasi"
        actor:
          type: string
          description: Person or agent, e.g. "S3 · Validasi" or "Budi (Checker)"
        detail:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...