# lib/api-zod/src/generated/types/reportUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LKPM-Flow API — companies, reports, data points, constraints, audit trail, and dashboards for LKPM (Laporan Kegiatan Penanaman Modal) reporting to OSS.
 * OpenAPI spec version: 0.1.0
 */
import type { PeriodType } from './periodType';
import type { ReportStatus } from './reportStatus';

export interface ReportUpdate {
  periodType?: PeriodType;
  /** @minLength 1 */
  periodLabel?: string;
  year?: number;
  deadline?: Date;
  status?: ReportStatus;
  narrative?: string;
  ossReceipt?: string;
  /** @nullable */
  makerName?: string | null;
  /** @nullable */
  checkerName?: string | null;
  /** @nullable */
  approverName?: string | null;
  /** @nullable */
  makerId?: string | null;
  /** @nullable */
  checkerId?: string | null;
  /** @nullable */
  approverId?: string | null;
}